refactor(api): search api endpoint
