+++ cbir/api/images.py
@@ -107,46 +107,3 @@ def remove_image(
             "index": index_name,
         }
     )
-
-
-@router.post("/images/retrieve")
-async def retrieve_image(
-    request: Request,
-    image: UploadFile,
-    nrt_neigh: int = Query(...),
-    storage_name: str = Query(..., alias="storage"),
-    index_name: str = Query(default="index", alias="index"),
-    retrieval: ImageRetrieval = Depends(get_retrieval),
-) -> JSONResponse:
-    """
-    Retrieve similar images from the database.
-
-    Args:
-        request (Request): The incoming HTTP request.
-        image (UploadFile): The query image.
-        nrt_neigh (int): The number of nearest neighbors to retrieve.
-        storage_name (str): The name of the storage where the index is stored.
-        index_name (str): The name of the index where the image features will be added.
-        retrieval (ImageRetrieval): The image retrieval object.
-
-    Returns:
-        JSONResponse: A JSON containing the list of filenames and their distances.
-    """
-
-    if not storage_name:
-        raise HTTPException(status_code=404, detail="Storage is required")
-
-    model = request.app.state.model
-    content = await image.read()
-
-    filenames, distances = retrieval.search(model, content, nrt_neigh)
-
-    return JSONResponse(
-        content={
-            "query": image.filename,
-            "storage": storage_name,
-            "index": index_name,
-            "filenames": filenames,
-            "distances": distances,
-        }
-    )
